remove/split AzurePlugin.java
