+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-lib-java/src/main/resources/META-INF/azure-intellij-plugin-lib-java.xml
@@ -7,4 +7,7 @@
                 class="com.microsoft.azure.toolkit.intellij.common.action.WhatsNewAction" text="What's New in Azure Toolkit"
                 description="Show new features of Azure Toolkit for Java"/>
     </actions>
+    <extensions defaultExtensionNs="com.intellij">
+        <postStartupActivity implementation="com.microsoft.azure.toolkit.intellij.common.action.WhatsNewStartupActivity"/>
+    </extensions>
 </idea-plugin>
