+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-sdk-reference-book/src/main/resources/META-INF/azure-sdk-reference-book.xml
@@ -4,6 +4,7 @@
         <dependencySupport displayName="Azure SDK client libraries(Track 2)" coordinate="com.azure:azure-core" kind="java"/>
         <dependencySupport displayName="Azure SDK client libraries" coordinate="com.microsoft.azure:azure-client-runtime" kind="java"/>
         <dependencySupport displayName="Library for Azure Java Functions" coordinate="com.microsoft.azure.functions:azure-functions-java-library" kind="java"/>
+        <postStartupActivity implementation="com.microsoft.azure.toolkit.intellij.azuresdk.ProjectSdkIntrospectionStartupActivity"/>
     </extensions>
 
     <extensions defaultExtensionNs="com.microsoft.tooling.msservices.intellij.azure">
