+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/main/java/com/microsoft/intellij/serviceexplorer/azure/SignInOutAction.java
@@ -7,7 +7,6 @@ package com.microsoft.intellij.serviceexplorer.azure;
 
 import com.intellij.openapi.project.Project;
 import com.microsoft.azure.toolkit.ide.common.icon.AzureIcon;
-import com.microsoft.azure.toolkit.intellij.AzurePlugin;
 import com.microsoft.azuretools.authmanage.IdeAzureAccount;
 import com.microsoft.azure.toolkit.intellij.common.auth.AzureSignInAction;
 import com.microsoft.tooling.msservices.serviceexplorer.NodeAction;
