+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/main/java/com/microsoft/intellij/ui/ActivityLogToolWindowFactory.java
@@ -16,10 +16,6 @@ import com.intellij.ui.table.TableView;
 import com.intellij.util.ui.ColumnInfo;
 import com.intellij.util.ui.ListTableModel;
 import com.intellij.util.ui.PlatformColors;
-import com.microsoft.azure.toolkit.lib.common.task.AzureTaskManager;
-import com.microsoft.azuretools.azurecommons.deploy.DeploymentEventArgs;
-import com.microsoft.azuretools.azurecommons.deploy.DeploymentEventListener;
-import com.microsoft.intellij.AzurePlugin;
 import com.microsoft.intellij.util.PluginUtil;
 import org.jetbrains.annotations.NotNull;
 
@@ -50,7 +46,7 @@ public class ActivityLogToolWindowFactory implements ToolWindowFactory {
 
     @Override
     public boolean isApplicable(@NotNull Project project) {
-        return !AzurePlugin.IS_ANDROID_STUDIO;
+        return false;
     }
 
     @Override
@@ -89,41 +85,6 @@ public class ActivityLogToolWindowFactory implements ToolWindowFactory {
     }
 
     public void registerDeploymentListener() {
-        AzurePlugin.addDeploymentEventListener(
-                new DeploymentEventListener() {
-
-                    @Override
-                    public void onDeploymentStep(final DeploymentEventArgs args) {
-                        // unique identifier for deployment
-                        String key = args.getId() + args.getStartTime().getTime();
-                        if (rows.containsKey(key)) {
-                            final DeploymentTableItem item = rows.get(key);
-                            AzureTaskManager.getInstance().runLater(() -> {
-                                item.progress = args.getDeployCompleteness();
-                                if (args.getDeployMessage().equalsIgnoreCase(message("runStatus"))) {
-                                    String html = String.format("%s%s%s%s", "  ", "<html><a href=\"" + args.getDeploymentURL() + "\">", message("runStatusVisible"), "</a></html>");
-                                    item.description = message("runStatusVisible");
-                                    item.link = args.getDeploymentURL();
-                                    if (!ToolWindowManager.getInstance(project).getToolWindow(ActivityLogToolWindowFactory.ACTIVITY_LOG_WINDOW).isVisible()) {
-                                        ToolWindowManager.getInstance(project).notifyByBalloon(ACTIVITY_LOG_WINDOW, MessageType.INFO, html, null,
-                                                new BrowserHyperlinkListener());
-                                    }
-                                } else {
-                                    item.description = args.getDeployMessage();
-                                }
-                                table.getListTableModel().fireTableDataChanged();
-                            });
-                        } else {
-                            final DeploymentTableItem item = new DeploymentTableItem(args.getId(), args.getDeployMessage(),
-                                    dateFormat.format(args.getStartTime()), args.getDeployCompleteness());
-                            rows.put(key, item);
-                            AzureTaskManager.getInstance().runLater(() -> table.getListTableModel().addRow(item));
-                        }
-                        final ToolWindow window = Objects.requireNonNull(ToolWindowManager.getInstance(project).getToolWindow(ACTIVITY_LOG_WINDOW));
-                        window.setAvailable(true);
-                        window.activate(null);
-                    }
-                });
     }
 
     private class ProgressBarRenderer implements TableCellRenderer {
