--- PluginsAndFeatures/azure-toolkit-for-intellij/src/main/java/com/microsoft/intellij/ui/libraries/ApplicationInsightsPanel.form
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.microsoft.intellij.ui.libraries.ApplicationInsightsPanel">
-  <grid id="27dc6" binding="rootPanel" layout-manager="GridLayoutManager" row-count="4" column-count="2" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
-    <margin top="0" left="0" bottom="0" right="0"/>
-    <constraints>
-      <xy x="20" y="20" width="506" height="400"/>
-    </constraints>
-    <properties/>
-    <border type="none"/>
-    <children>
-      <component id="f30c0" class="javax.swing.JCheckBox" binding="aiCheck">
-        <constraints>
-          <grid row="0" column="0" row-span="1" col-span="2" vsize-policy="0" hsize-policy="3" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <text resource-bundle="com/microsoft/intellij/ui/messages/messages" key="aiCheckBoxTxt"/>
-        </properties>
-      </component>
-      <vspacer id="ca7b">
-        <constraints>
-          <grid row="3" column="0" row-span="1" col-span="2" vsize-policy="6" hsize-policy="1" anchor="0" fill="2" indent="0" use-parent-layout="false"/>
-        </constraints>
-      </vspacer>
-      <component id="e7a2d" class="javax.swing.JLabel" binding="lblInstrumentationKey">
-        <constraints>
-          <grid row="1" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <text resource-bundle="com/microsoft/intellij/ui/messages/messages" key="lblInstrumentationKey"/>
-        </properties>
-      </component>
-      <component id="3d146" class="org.jdesktop.swingx.JXHyperlink" binding="lnkAIPrivacy">
-        <constraints>
-          <grid row="2" column="1" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="4" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <text resource-bundle="com/microsoft/intellij/ui/messages/messages" key="AIPrivacy"/>
-        </properties>
-      </component>
-      <component id="ad696" class="com.microsoft.azure.toolkit.intellij.common.AzureComboBox" binding="comboInstrumentation" custom-create="true">
-        <constraints>
-          <grid row="1" column="1" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
-        </constraints>
-      </component>
-    </children>
-  </grid>
-</form>
